+++ processor/src/test/resources/expected/com/example/autorun/Arez_ReadWriteAutorunModel.java
@@ -15,10 +15,10 @@ import javax.annotation.Nullable;
 import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
-public final class Arez_ReadWriteAutorunModel extends ReadWriteAutorunModel implements Disposable, Identifiable<Long>, ComponentObservable {
-  private static volatile long $$arezi$$_nextId;
+public final class Arez_ReadWriteAutorunModel extends ReadWriteAutorunModel implements Disposable, Identifiable<Integer>, ComponentObservable {
+  private static volatile int $$arezi$$_nextId;
 
-  private final long $$arezi$$_id;
+  private final int $$arezi$$_id;
 
   private byte $$arezi$$_state;
 
@@ -35,7 +35,7 @@ public final class Arez_ReadWriteAutorunModel extends ReadWriteAutorunModel impl
   public Arez_ReadWriteAutorunModel() {
     super();
     this.$$arezi$$_context = Arez.areZonesEnabled() ? Arez.context() : null;
-    this.$$arezi$$_id = ( Arez.areNamesEnabled() || Arez.areNativeComponentsEnabled() ) ? $$arezi$$_nextId++ : 0L;
+    this.$$arezi$$_id = ( Arez.areNamesEnabled() || Arez.areNativeComponentsEnabled() ) ? $$arezi$$_nextId++ : 0;
     this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
     this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().createComponent( "ReadWriteAutorunModel", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null ) : null;
     this.$$arezi$$_disposedObservable = $$arezi$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".isDisposed" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arezi$$_state >= 0 : null );
@@ -56,7 +56,7 @@ public final class Arez_ReadWriteAutorunModel extends ReadWriteAutorunModel impl
     return Arez.areZonesEnabled() ? this.$$arezi$$_context : Arez.context();
   }
 
-  final long $$arezi$$_id() {
+  final int $$arezi$$_id() {
     if ( Arez.shouldCheckInvariants() && !Arez.areNamesEnabled() && !Arez.areNativeComponentsEnabled() ) {
       Guards.fail( () -> "Method invoked to access id when id not expected." );
     }
@@ -65,7 +65,7 @@ public final class Arez_ReadWriteAutorunModel extends ReadWriteAutorunModel impl
 
   @Override
   @Nonnull
-  public final Long getArezId() {
+  public final Integer getArezId() {
     return $$arezi$$_id();
   }
 
@@ -121,7 +121,7 @@ public final class Arez_ReadWriteAutorunModel extends ReadWriteAutorunModel impl
   @Override
   public final int hashCode() {
     if ( Arez.areNativeComponentsEnabled() ) {
-      return Long.hashCode( $$arezi$$_id() );
+      return Double.hashCode( $$arezi$$_id() );
     } else {
       return super.hashCode();
     }
